requirements.txt
# quant_project/requirements.txt
# 基础依赖 (开源版)
akshare>=1.11.0
baostock>=0.2.0
backtrader>=1.9.0
pandas>=2.0.0
numpy>=1.26.0

# 付费版专属依赖
# 推送功能
pyyaml>=6.0
requests>=2.31.0

# ML/AI功能
scikit-learn>=1.3.0
scipy>=1.11.0

# Web界面依赖
streamlit>=1.28.0
plotly>=5.18.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# 代码质量
flake8>=6.1.0
black>=23.9.0
mypy>=1.5.0
Pillow>=10.0.0
